pyproject.toml
[project]
name = "text-to-sql-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.16",
    "faiss-cpu>=1.12.0",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "pdf2image>=1.17.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
